# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAPI Laravel API",
        "description": "API REST que integra la Star Wars API (SWAPI) con caché, reintentos y DTOs. Gestiona planetas y habitantes localmente con relaciones Eloquent.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API v1 – local"
        }
    ],
    "paths": {
...
        "/planetas": {
            "get": {
                "tags": [
                    "Planetas Locales"
                ],
                "summary": "Listar planetas locales con filtros",
                "operationId": "ccbeda733cdde1756273d3563c43aa28",
                "responses": {
                    "200": {
                        "description": "Lista paginada de planetas"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Planetas Locales"
                ],
                "summary": "Crear un planeta con sus habitantes",
                "operationId": "54a10177712cadffb397b9ca14dea71a",
                "responses": {
                    "201": {
                        "description": "Planeta creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/planetas/{id}": {
            "get": {
                "tags": [
                    "Planetas Locales"
                ],
                "summary": "Obtener un planeta local con sus habitantes",
                "operationId": "7135fa518db4d6e6103cf26e2c046d33",
                "responses": {
                    "200": {
                        "description": "Planeta encontrado"
                    },
                    "404": {
                        "description": "Planeta no encontrado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Planetas Locales"
                ],
                "summary": "Actualizar un planeta local",
                "operationId": "564274b3132ca6343c243312029359be",
                "responses": {
                    "200": {
                        "description": "Planeta actualizado"
                    },
                    "404": {
                        "description": "Planeta no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Planetas Locales"
                ],
                "summary": "Eliminar un planeta local",
                "operationId": "8c556472b5de6c482e28c4852638758a",
                "responses": {
                    "200": {
                        "description": "Planeta eliminado"
                    },
                    "404": {
                        "description": "Planeta no encontrado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key requerida. Configura API_KEY en tu .env y envíala como header X-API-KEY.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}